pyproject.toml
[project]
name = "okx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "python-okx>=0.4.0",
    "requests>=2.32.5",
]
